back-end/requirements.txt
# ============================================
# Portfolio Backend - Dependencies
# ============================================
# Clean Architecture + Hexagonal (Ports & Adapters)
# ============================================

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
pymysql==1.1.0
cryptography==42.0.0

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
slowapi==0.1.9  # Rate limiting

# HTTP Client (for external APIs)
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
python-dotenv==1.0.0

# Optional: PostgreSQL (for production)
# psycopg2-binary==2.9.9

# Optional: Redis (for caching)
# redis==5.0.1

# Optional: Celery (for background tasks)
# celery==5.3.6
